feat: Add lineErrorf
