+++ tools/schema/model/schema_types.go
@@ -1,10 +1,6 @@
 package model
 
-import (
-	"time"
-
-	"github.com/iotaledger/wasp/packages/iscp"
-)
+import "fmt"
 
 // TODO describe schema details in docs
 type (
@@ -55,37 +51,6 @@ func NewSchemaDef() *SchemaDef {
 	return def
 }
 
-type Func struct {
-	Name    string
-	Access  DefElt
-	Kind    string
-	Hname   iscp.Hname
-	Params  []*Field
-	Results []*Field
-	Line    int
-	Comment string
-}
-
-type Struct struct {
-	Name   DefElt
-	Fields []*Field
-}
-
-type Schema struct {
-	ContractName  string
-	PackageName   string
-	Description   string
-	CoreContracts bool
-	SchemaTime    time.Time
-	Events        []*Struct
-	Funcs         []*Func
-	Params        []*Field
-	Results       []*Field
-	StateVars     []*Field
-	Structs       []*Struct
-	Typedefs      []*Field
-}
-
-func NewSchema() *Schema {
-	return &Schema{}
+func LineErrorf(lineNr int, format string, args ...interface{}) error {
+	return fmt.Errorf("line %d: %s", lineNr, fmt.Sprintf(format, args...))
 }
